requirements.txt
asgiref==3.8.1
Deprecated==1.2.14
Django==5.0.6
image==1.5.33
img2pdf==0.5.1
lxml==5.2.2
packaging==24.0
pikepdf==8.15.1
pillow==10.3.0
pypdf==4.2.0
six==1.16.0
sqlparse==0.5.0
wrapt==1.16.0
